fix: e2e tests setup
